Add margin_bottom param to Attachment component
The Attachment component defaults to having no margin-bottom.

However sometimes we need it to have a margin-bottom value, for example
when rendering a list of attachments on a document. Otherwise the
components look too 'squashed up'.

This has previously been worked around by wrapping the component in a
container element with margin bottom. [1] But this feels a bit hacky and
becomes trickier when we don't have so much control over the way the
markup around the component (for example, when it's been embedded within
a Govspeak document).

[1]: https://github.com/alphagov/government-frontend/blob/0bec9a842902db2a7353d28400fa7bf5d7008c8b/app/views/content_items/_attachments.html.erb#L18-L23