Wrap the text of start buttons in a span
When using an inline element (such as `<abbr>`, `<address>` or
`<telephone>`) in a link with the `govuk-button--start` class applied to
it, there is no spacing between the inline element and the text of the
link.

The reason that this is occurring is the `display: inline-flex;` rule in
the `govuk-button--start` class.

The suggested solution was to wrap the text of `start` buttons in a
`<span>` tag when rendering a start button:
https://github.com/alphagov/govspeak/pull/277#issuecomment-1620327662

This is the solution I'm implementing here.