Remove old technique for adding tracking to attachment links
Previously in whitehall we would just add one data attribute to the attachment links. We would then apply the GA4 attributes from here using JS. This was because changes to the Whitehall HTML attachments were expensive, it required a republish of 45,000 documents. So the idea was to add a data-attribute once, and then in publishing components control the actual GA4 attributes so they were easier to change. This is no longer needed, since Whitehall now uses the Attachment component from this repo.