Render search button & input high contrast outline
Add a transparent outline to the search button in all modes, and to form inputs in the blue background mode. This way Windows high contrast mode renders an outline for the elements, which helps to distinguish them from their background and makes them easier to identify. This does not result in any visual changes in the standard contrast mode.

This is to fix an issue which results from the removal of the default border from the search button in all modes, and the default border from from form input elements in the blue background mode, in order to comply with the GOV.UK design. This meant that these elements didn't have a visible outline in Windows high contrast mode which could make them difficult to distinguish from their background, or to identify. (This wasn't an issue in Firefox when users changes the default browser colours there because Firefox renders a distinct colour background inside buttons and form inputs.)