Fix row spacing on topics browse pages
The existing grid setup forced every row to be the same height as the
tallest row (content dependent). But this causes large gaps of
whitespace if one content item is larger than all the others, and a
peculiar behaviour for the first row where it would always collapse,
leaving row heights inconsistent.

Removing `grid-template-rows: auto` prevents the first row collapsing
and removes the whitespace anomalies for the other rows.

Note that this change also requires guidance to be provided for
content authors to ensure that each content item is of similar length
which helps keeps the row heights consistent.