Fix bug where all links under an attachment were tracked the same
As we put the JSON on the parent of the attachment, every link was being tracked the same. However there are email links and 'Order a copy' links that we don't want tracked with our JSON attributes. Therefore we move the JSON to only be on the links we want to track with those attributes.