Add '+' as character to exclude in PII email regex
Without this, a search term in the URL containing a URL gets completely replaced with [email], as it counts characters before and after a + symbol as part of the email address. For example, searching hello+world+email@example.com+SW12AA+1990-01-01 would just return [email] with the previous regex. One issue with this solution however is that + is a valid character in gmail. So if someone entered their email as first.lastname+test@gmail.com then our PII would not redact it properly