Split data-ga4-section into two attributes so that sections can be set on individual filter inputs
Previously, data-ga4-section was used to set indexes on filter parent elements, as well as give a label to the filter. However, there has been a request for the labels to be more granual. For example, instead of having one label Topics, have a label Topic and Sub-topic on the individual filters, rather than one label on their parent div. To achieve this, data-ga4-section is now used to set the label on each input. data-ga4-filter-parent now exists in order to identify where to calculate the indexes for each filter.