WIP: Add ga4-index data attrs in Ruby vs JS
We're now able to provide the index from the facet partials thus
avoiding the need for the code in
GOVUK.analyticsGa4.Ga4FinderTracker.setFilterIndexes(). However, I've
left the code to trigger the "ga4-filter-indexes-added" event for now to
reduce the scope of this change. The plan is that the whole of
setFilterIndexes() will become redundant in a subsequent commit.

Note that the index is 0-indexed vs 1-indexed since the former is how
the Rails indexing was working. This means we currently have to add 1 in
the facet partials when setting the ga4-index.index_section value. It
might make more sense to change the index in FacetsIterator to be
1-indexed and change OptionSelectFacet#closed_by_default? to work with
this.

TODO: I'm not completely convinced that the Facet#has_ga4_section?
method belongs in the model.