Fix the HTML language never being used to set GA4 language
If the #content element existed, but had no language on it, it would return this.nullValue instead of jumping out to check the html language value instead. This has been fixed, so that it wont return null when content doesnt have a language on it. It will now only return null if both #content and html do not have a language on them