Support input has_error true w/o error_message/error_items
It doesn't make sense to me to render an empty error-message paragraph
element if no error_message or error_items are set. However, it does
still make sense to add the error CSS class to the input element.

This commit changes the input component so it only renders the
error-message paragraph if error_message is present or there are some
error_items.

My motivation for making this change is so that I can change the
date-input component to support errors on individual input elements,
e.g. just the year input element. This usage is described in the docs
for the date-input component [1]:

> If you’re highlighting just one field - either the day, month or year
> - only style the field that has an error. The error message must say
> which field has an error, like this:

However, highlighting just one field like this is not actually currently
supported by the date-input component which is a bit confusing!

[1]: https://design-system.service.gov.uk/components/date-input/#error-messages