Update image-card two thirds size attributes
Set the correct width and height attributes when using the new two_thirds variation of the image-card component.

The aspect ratio for the featured image size to be used on the new homepage design is 1:1 when using the new two_thirds variation of the image card component.

Add related tests to ensure that the correct width and height attributes are set.