More changes to the navbar
Change gutter of navbar on homepage so it is 24px like the gutter in the
homepage. Force font sizes of headers in Search dropdown menu to be 24px
at all screen sizes. Make padding of the search drop menu 50px in total.
Make font sizes of the link description in the navbar 19px and update
margin-top for large navbar.