Add separate manifest for sassc-rails compatibility
The default manifest `govuk_publishing_components_manifest.js`,which adds a dependency on the builds folder, is used if `dartsass-rails` is bundled. Otherwise, `govuk_publishing_components_sassc-rails_manifest.js`, which adds dependencies on individual component stylesheets, is used if `sassc-rails` is bundled.
Also, include a test to check the use of the correct manifest.

Co-Authored-By: Martin Jones <28779939+MartinJJones@users.noreply.github.com>