Make global relative links absolute
There is an issue with CSV preview pages as they are hosted on a different domain to www.gov.uk - this breaks all the footer and header links as they are relative, so when clicking them they 404 as the domain is not www.gov.uk. The search 'action' and feedback form 'action' are also broken on this other domain as they are relative. Therefore this code takes the Plek.website_root or GOVUK docker VIRTUAL_HOST, and appends it to the start of relative URLs so that they are always absolute, preventing our links from breaking on non www.gov.uk domains.