Update initialisation script snippet in layouts
For govuk-frontend v5, there now needs to be additional class added to
the `body` tag to indicate that the components are supported. This class
is `govuk-frontend-supported` and is added if the browser supports JS of
type `module` (which is the new browser target for govuk-frontend). The
script tag is directly taken from the documentation.