Use govuk-frontend v5 component initialisation
In govuk-frontend v5, initialisation has been moved to creation of
component JS module. This means that `init()` does not need to be
called to initialise a component, it is instead called in the
constructor.

To ensure that modules from govuk-frontend and modules from the gem
can be initialised easily in the same `initAll()` method, the modules
of the gem have been changed to use the new initialisation method. This
means moving `init()` to the constructor of each module. It also means
updating the `initAll()` method to not call 'init()' of each module.