Use govuk-frontend v5 component initilization
In govuk-frontend v5, initilization has been moved to creation of
component JS module. This means that `init()` does not need to be
called to initialize a component, it is instead called in the
constructor.

To ensure that modules from govuk-frontend and modules from the gem
can be initialized easily in the same `initAll()` method, the modules
of the gem have been changed to use the new initilization method. This
means moving `init()` to the constructor of each module. It also means
updating the `initAll()` method to not call 'init()' of each module.