Update tag component in light of govuk-frontend v5
govuk-frontend v5 introduces some changes to the tag component. The one
that has the most impact, is that the tag is now designed not to use
ALL CAPS text anymore. Instead it is to be titlecase. In addition there
are new background colours introduced which means that the styling for
some tags in govuk_publishing_components has to be changed to maintain
sufficient contrast. Tests have been updated to anticipate titlecase
tag text instead of ALL CAPS text.