Use Terser instead of Uglifier
govuk-frontend v5 now targets browsers that support ES6.

This means that the UMD modules used in govuk_publishing_components from govuk-frontend
use features of ES6 and so it means that Uglifier can't be used anymore because it only supports ES5.

As well as installing Terser and updating the config, this commit also contains a patch for getting Terser working. Sprockets doesn't have an built-in loader for Terser so we need to add this functionality.