Separate ES6 components from ES5 components
Add new config to state if application is loading ES6 components in 'es6-components.js'. If set to true `layout_for_public` and `layout_for_admin` will include a script tag with type `module` that points to `es6-components.js`. If this config is set to true and there is no `es6-components.js` then an error will occur.