Audit now checks if component uses govuk-frontend
If a component being used in an application is uses govuk-frontend
ES6 JS, then a tag will be added to indicate this in the Component
Audit view. This is useful to determine which components are using
ES6 JS and need to have their JS be loaded in a different file.