Create new political meta tags for GA4
Some pages have their 'government' object in the content item nested, so the existing political meta tags would not render on these pages. This impacted the GA4 tracking. Therefore, we create two GA4 meta tags which account for the 'government' object being in a different place in the content item.