Fix YouTube image-card layout
The YouTube card image overlapped with the text, removing the "tricks" fixes the rendering issue and no longer appears to be required.

The approach used here was originally used to help improve keyboard navigation when tabbing into the player, however this no longer appears to be an issue with the YouTube player on desktop screen sizes.