Ensure nested links with their own data-ga4-link will still work
If the parent element has data-ga4-limit-to-element-class, but the link itself has its own object too, the link tracker was firing twice as it would use the parent elements GA4 JSON, and the GA4 JSON on the link itself. This will prevent that from happening, by ensuring the closest ga4-link JSON object is used for tracking.