Mainstream browse breadcrumbs
When a piece of content is tagged to mainstream browse, and has a mainstream
browse parent, we want to add a mainstream browse breadcrumb. Content like that
occurs when mainstream content is published in Publisher, and has a mainstream
browse breadcrumb selected in the UI.

If a piece of content is tagged to mainstream browse, but does not have a mainstream
browse parent, we want it to have the default taxonomy topic breadcrumb. This situation
might occur when Whitehall content is tagged to MSB in content tagger. The current
agreed pattern for crumbs across GOV.UK is that Whitehall content should have a taxonomy
topic breadcrumb.