Fix GA4 specialist link tracker when its inside the other link tracker
When the other link tracker is on a parent element, but is limiting itself to certain classes, all the specialist links were not being tracked. This fixes that.

Also fixes a flaky test.

Whenever GA4 is initialised in tests, the specialist link tracker is initialised. Therefore tracking on it needs to be disabled after each time analytics is initialised. This was done in certain tests, however the tracking was still being duplicated somewhere that I could not find. Therefore, it's easier and safer going forward to just stop the tracking after every test at the highest level instead of trying to figure out each individual place it is being initialised.