Increase font size and image size in organisational logos
The original treatment for logos used multiple sizes and screen density media queries to send low/high resolution images depending on the screen density. By increasing the size of the logos, the low resolution images are now too small so support for those has been dropped and we now use only the larger sizes. These larger images are scaled down by the browser and so will look sharper enough on low density screens.

The mixin for the crest was also changed to allow the passing of different offsets, as some logos now need small adjustments to optically match the alignment with the text.