Fix rendering of unnumbered papers
The "false" check in the unnumbered command and act papers was failing
because the component is passed an empty string rather than nil for the
number, which does not evaluate to false:

e.g.

a = false
!a
=> true
b = nil
!b
=> true
c = ""
!c
=> false

Therefore the check was returning `nil` rather than the "Unnumbered..."
text.