Add favicon for test environment
When we run our system specs that use the layout_for_admin component they
fail due to this line of code

```
<%= favicon_link_tag "govuk_publishing_components/favicon-#{environment}.png" %>
```

This is because the `favicon-test.png` file does not exist. There are favicnons
for:

- prod
- staging
- integration
- example (component guide)
- development

I've duplicated the development favicon and named it favicon-test.png
to unblock our system specs.