Remove email from survey
What

Replace the "email me a survey link" form when the user clicks "No" on
"is this page useful" with a simple link.

Having a public facing form which takes email addresses and sends them
emails is a potential vector for abuse.

We think that this functionality was originally added because we'd
prefer users to fill in the survey after they've completed their task on
GOV.UK, so we get a more accurate picture of their intent and
experience. Having the survey in their inbox is less intrusive, and
doesn't take them away from GOV.UK.

However, in hindsight, having an email step feels like an unecessary bit
of friction compared to just giving the user the link (which we already
did, in the "Don't have an email address?" link).