WIP - Switch from standardx to eslint with standard config
[standardx](https://github.com/standard/standardx) appears to be a
bit... unmaintained. It's pinning us to a pretty old version of eslint,
and a few other dependencies.

We switched to using eslint in Whitehall in https://github.com/alphagov/whitehall/pull/8178

We should consider rolling this out everywhere.

TODO - more details on what dependencies we're stuck on, and the
horrible story of the eslint config file format :see_no_evil: