Update checkbox display from `block` to `flex`
In V5 of the design system, checkboxes use flexbox, `display: flex`.

In the option select component, visibility of the checkbox is set using either `display: none` or `display: block`. Setting the value to `block` caused rendering issues when focusing a checkbox item. Setting the value to `flex` aligns with the approach used in the design system and fixes the rendering issue.