Fix focus styling in super navigation header logo
The focus style of the super navigation header logo was not rendering correctly in the super navigation header menu, but it was styled correct in the layout-header component.

To fix this, the line height is set to 1 in `.gem-c-layout-super-navigation-header__header-logo`, this matches the approach used for `.govuk-header` in the layout-header from the design system.