Check govuk-frontend accordion has initialised
When the accordion component from govuk-frontend is initialised, the `govuk-accordion__show-all` class is added to the button element.

It is possible that the accordion component from govuk-frontend does not get initialised, but we have no checks in place to guard against this when initialising the accordion component from the gem.

This change checks if the accordion component from govuk-frontend has initialised before trying to initialise the accordion component from the publishing components gem.

A new test has been added to check it does not initialise, if the accordion from govuk-frontend has not initialised. `gem-c-accordion__show-all` was removed from the fixture as this is added by the JS from the publishing-components accordion.