Only set gtm_cookies_win=x for preview in GTM script
The GTA doesn’t work as expected for DGU.

It appears that gtm_cookies_win=x allows GTM Preview to change the environment
of a page, which means that they win over gtm_auth and gtm_preview parameters.

Setting gtm_cookies_win=x  let the GTM editor decide which environment they
want to use, regardless of what was in the tracking snippet. 

This makes it identically to the behaviour in ga4-core.js. The commit message
suggest that it can slow down GTM:
https://github.com/alphagov/govuk_publishing_components/blob/eb66a18f86/app/assets/javascripts/govuk_publishing_components/analytics-ga4/ga4-core.js#L25

https://govuk.zendesk.com/agent/tickets/5626599