Disable browser corrections for header search field
This applies the `disable_corrections` setting from #4112 to the
`search` components used in the `layout_super_navigation_header` and
`layout_header` components.

We expect this to improve the quality of search results for users on
mobile as their browsers will no longer excessively autocorrect, and in
case of genuine misspellings, the site search engine can cope with that
well anyway.