Resolve DartSass mixed declaration deprecations
Sass has deprecated mixed declarations [1] and this generates us
hundreds of warnings for versions of dartsass with this deprecation [2]
which we see in this project after the update to dartsass-rails 0.51
[3].

This commit resolves these deprecations and when combined with Andy's PR
[4] all Sass warnings are resolved.

This mixed declaration warning is super annoying for us. In short it
prevents mixing declarations on an element with those of a nested
element and requires the ones on the individual element to come first.
This seems fine in theory however in practice we use mixins as a way to
set nested declarations (generally media query) at the same time as
setting an element declaration - e.g. the govuk-font mixin.

The way to resolve this has been to tediously move the use of mixins
that have nested rules so that they occur after the element rules. It'll
be very easy to re-introduce warnings here so it'll be good when
DartSass turn the old behaviour off.

The risk on the changes made here is that some of the CSS that was
following the rules before overrode declarations from the mixin and thus
creates a visual difference. Although I've tried to be careful I am
somewhat relying on Percy to catch issues here.

The only visual difference identified is the line-height of
.gem-c-organisation-logo in the organisation-logo. Previously the
line-height declaration of 1.35 was overriden by the media query for the
font declaration for screen sizes of 40.0625em to be set to
1.3157894737. I expect that this was a mistake given we didn't have a
media query in the declaration and the new value - 1.35 - is very
similar.

[1]: https://sass-lang.com/documentation/breaking-changes/mixed-decls/
[2]: https://github.com/alphagov/govuk_publishing_components/actions/runs/10181528374/job/28161909452?pr=4123
[3]: https://github.com/alphagov/govuk_publishing_components/pull/4123
[4]: https://github.com/alphagov/govuk_publishing_components/pull/4106