Drop support for Ruby 3.1
Ruby 3.1 is a few months from end-of-life and in security fixes only
territory. This update is somewhat forced upon us by the sass-embedded
dependency which has dropped support for Ruby 3.1 in some circumstances
[1] which causes the dartsass-rails gem to fail [2].

Although we could probably find a way to hold the version of
dartsass-rails back for Ruby 3.1, we probably shouldn't as we'll get a
similar experience of apps using this gem if we maintain similar
versions.

The consequences of this change should be pretty much negligible as
there are not GOV.UK apps on Ruby 3.1 anymore and the .ruby-version for
this repo was already (yet mistakenly) updated to Ruby 3.2.

[1]: https://github.com/sass-contrib/sass-embedded-host-ruby/blob/19323666b359bf3d58b562af09fdc75440453608/sass-embedded.gemspec#L39-L43
[2]: https://github.com/alphagov/govuk_publishing_components/pull/4123