Fix the search toggle in layout header
On mobile and tablet screen sizes clicking on the search icon in the layout header did not reveal the hidden search form element, `gem-c-layout-header__search-form`.

The reason for this is that as of V3.0.0 of govuk-frontend, classes have a `govuk-` prefix. The HTML for the search template has been updated to reflect this, as well as using the `aria-controls` attribute to target the element to be toggled.

The toggle functionality, makes use of the `hidden` attribute to show/hide the target element, `display: none` has been removed from `.gem-c-layout-header__search-form` to ensure that the search form visibility can be toggled and can be accessed when JS is disabled.