Fix print issues on layout-super-navigation-header
- Background should always print transparent
- Foreground should always print black
- Remove top border
- Add bottom border to match new changes in cross service header
- Remove left/right margins not required for print

Note the use of the :has selector to remove the border and add a left indent
only if the component contains the hidden 'GOV.UK' text. Variants that have
hidden logo text are only used on the homepage, and need an adjusted design
for print so that the crown logo lines up with the intro text below.