Fix issue with layout header in component guide
When previewing all variations of the layout_header component in the component guide, only the first instance of the navigation menu and the search would be toggled. Clicking on the "Show or hide Top Level Navigation" in the 2nd instance, would toggle the Top Level Navigation for the 1st.

The fix used here is to give each instance a unique id to ensure the toggle functionality will target the correct element.

Although it seems unlikely that we will have more than 1 instance of a layout header on a page, other than the component guide, this pattern of setting a unique ID is commonly used in other components, `checkboxes` for example and ensure the component guide will work as expected.