Remove aria-label on menu button
As part of a DAC audit of GOV.UK Chat where this component is used, it was noted that the difference in the aria-label and the actual text of the menu button constituted a failure of SC 2.5.3 Label in Name (A). DAC's recommendation was to remove the aria-label and to just use the button text.