Document list: parts should respect `remove_underline`
The links to parts/"subdocuments" on the document list component are
currently always rendered without underlines, regardless of whether
`remove_underline` is actually set on the component itself. This is
probably because it's only used in Finder Frontend's search results
page, which always uses `remove_underline`.

We are starting to implement a UI overhaul of the search results page
which has underlines on all links in the document list component,
including those for parts.

- Make links to parts render with underlines unless `remove_underline`
  is set on the component
- Remove custom CSS for removing underlines when `remove_underline` is
  set, and just use `govuk-link--no-underline` instead