Use component wrapper in hint component
This adds the component wrapper to the hint component in order to
provide additional options to a hint.

The primary motivation of this was that we want to dynamically change
the value of a hint in GOV.UK Chat and thus want to set a .js-class to
act as a selector [1] (private repo) and instead are using a custom id
to work around it.

Looking at other components it seemed that using the component wrapper
was a more conventional approach than just utilising the shared helper
to add only classes.

[1]: https://github.com/alphagov/govuk-chat/pull/660#discussion_r1750067107