Work around `accessible-autocomplete` callback order
The accessible-autocomplete component calls the provided `onConfirm`
callback which we use to submit the containing form _before_ it updates
its internal state, so the value of the input field is not yet updated
at that point.

This works around that issue by setting the value of the input field
manually at the point before we submit the form (we only need the form
field to be set, the rest of the component state is okay to catch up
later and will be moot anyway due to page reload).