Add ostruct as an explicit dependency
Ruby is dropping ostruct from standard library and will require it to be
installed as a gem. When using an app with Ruby 3.3.5 there is a
warning:

```
/root/.rbenv/versions/3.3.5/lib/ruby/3.3.0/bundled_gems.rb:75: warning: /root/.rbenv/versions/3.3.5/lib/ruby/3.3.0/ostruct.rb was loaded from the standard library, but will no longer be part of the default gems starting from Ruby 3.5.0.
You can add ostruct to your Gemfile or gemspec to silence this warning.
Also please contact the author of govuk_publishing_components-43.3.0 to request adding ostruct into its gemspec.
```