Center text and icons vertically in the option select component
The font size of this element increased with the typography changes
deployed in a recent update to `govuk-frontend`. This means that the
up/down icons no longer align vertically. This PR centers the icons
with the text for better vertical balance. A new container for the
icon and button is also required so that the selected counter element
can be styled independently.

This also supports the same change made in `finder-frontend` where
search filter titles had the same tweak to icon positions. See the PR
at https://github.com/alphagov/finder-frontend/pull/3334.