Use breadcrumbs based on organisation if no parent
Use breadcrumbs based on ancestors, if the document has a linked parent.
Otherwise use breadcrumbs based on the first normal
(non-world) organisation the page is linked to.
If there are no linked organisations use taxon based breadcrumbs,
or if taxons are also not linked to,
the default breadcrumbs pointing to the homepage.