Show all big_number symbol types in our docs
We were only showing one example of the symbols that can be used in our docs. This led to people using inspect element to change the symbol on the docs page to test CSS changes. The styles are actually slightly different based on which symbol is passed to the component - therefore using inspect element isn't an accurate representation of what the symbol will look like when rendered. Therefore we should show extra symbol examples so that people aren't using inspect element to visualise other symbol types.