Update organisation logos
The logos organisation have been updated to reflect the most up-to-date organisation crests.

Images that previously had a 52px height have been updated to 64px, the reason for this is that the logo height is set to 32px in the CSS, using 64px as the height will help ensure we serve a 2x image as intended. Images that are 68px in height use a max-height of 34px, so no change in dimensions was made here.

Update portcullis_18px_x2.png