Show components not used by app in comp guide, instead of all gem comps
We now render a list of all gem components only in the gem component guide itself. In app component guides, we show "Gem components not used by this app" instead. The "All gem components' heading" can be removed as it was never used by the gem component guide. Change "unless" syntax to "if not" as we now have a an "else" condition.

Co-authored-by: Rebecca Pearce <17481621+beccapearce@users.noreply.github.com>
Co-authored-by: Hanna Laakso <hanna.laakso@digital.cabinet-office.gov.uk>