Layout super nav header: Correct name for search input
Sets an explicit field name (`keywords`) for the search input on the
layout super navigation header, as this is coming through into Finder
Frontend as the default name of `q`.

Finder Frontend implicitly converts this param so it's not actually
causing problems for users, but it's causing issues with our analytics
tracking.