Run Percy in main component guide
Percy uses the list of all gem components to take screenshots. Since the previous commit moved this list to the gem or main component guide, Percy was failing.

Set an environment variable to make Percy run in the main component guide.

This makes the tests pass, however I'm not sure if this is the approach we should be taking or if we should configure Percy differently. I'll chat with the team to decide if doing this here is out of scope and and if we should just render a list of all gem components in all views for now.