Add new custom analytics tracker for search
This adds a new `Ga4SearchTracker` analytics tracker which will
eventually replace:
- the usage of the generic `Ga4FormTracker` on the homepage and layout
  super navigation
- the custom search event code on the all content finder ("site search")

For now, this fires exactly the same event with the same data as the
existing use of the form tracker/custom search event code. However, it
allows us to share logic and be consistent in event data between all the
different places we offer a search box across GOV.UK, and will allow us
in a future change to add the ability to track usage of the upcoming
autocomplete functionality as well.