Add a content block component
Currently, the logic for rendering content blocks logic is locked up in
the Publishing API. The Content Modelling team has need to use this
logic in multiple places outside the Publishing API, so rather than have
to create an API endpoint to render a content block, we decided we
should have one source of truth.

Once this is in, we’ll call this component when rendering content blocks
in Publishing API, so the Publishing Components gem becomes the source
of truth.

Currently some of the blocks look a little bare bones, but we will be
expanding on this in the near future, especially when it comes to the
contact block.