Initial implementation of "add another" component
This is more or less a lift and shift of the existing "Add another" component from Whitehall Publisher. The Whitehall component in turn was inspired by the MoJ's "Add Another" component.

The component takes a fragment based approach, which means that the consumer is responsible for rendering the buttons for removing a set of fields from their form. This decision was made so that the buttons can be present as part of the HTML markup. The MoJ component creates the remove button within Javascript, which means it requires JS to be enabled to function. Making the user responsible for rendering the removal button means that they can use it as a form submit button in cases where JS is not enabled, and re-render the page without the specified field set. When JS is enabled, the module overrides the types of the buttons to "button" so that form submissions do not take place.