Only apply the tracker to the `ul` in tabs
Adding `ga4-event-tracker` on the container element meant that any tracking would apply not just to the tabs, but also the content of any of the tabs - any component that had tracking built in such as details would result in double tracking.
This behaviour was confusing and could easily result in difficult to spot issues.