Fix `search_with_autocomplete` double submit scenario
Our workaround for the `accessible-autocomplete` library hogging the
Enter key when the menu is open was too greedy, and also submitted the
form when the menu in fact wasn't open.

This adds a check to the workaround to ensure the menu is actually
shown, as the behaviour doesn't actually happen otherwise.