Fix "sticky" GA4 search tracker `tool_name`
The GA4 search tracker was only setting the `tool_name` on its `search`
event in case the user had accepted an autocomplete suggestion. Now that
we persist the `tool_name` in the page's analytics state, this meant
that if the user had accepted a suggestion on a previous search, the
`tool_name` would continue to be set for subsequent searches regardless
of whether they've actually _used_ autocomplete.