search_with_autocomplete: Scroll component into view on small screens
This adds a focus event listener on the search field to scroll the
viewport on small screens when the user focuses the search field so that
the component (including label) is on the top of the viewport.

This makes sure users are able to see the suggestions offered below the
input field and also avoids them potentially being obscured by the
keyboard.

Also adds a missing CHANGELOG entry from a previous PR (sorry!)