Allow licence details in footer to be hidden
The layout footer component is used quite a lot by internal services,
where it doesn’t make sense to show licence details. This adds a
`show_licence` option to allow users to hide the licence if they
want/need to. This will always default to `true`, unless the option
is specifically set to `false`.