Stop setting `id` on fieldsets
Now that the fieldset component uses the ComponentWrapperHelper, `id` should be handed by that instead of being explicitly called out.

This will result in `id` no longer being set on the fieldset, but instead on the parent. Jump links, the original usage for `id` will be unaffected, but any applications using `id` for other reasons (such as tests) may need to be updated.