[WIP] Allow search autocomplete to be easily disabled
Autocomplete uses a denylist to stop certain terms from serving
suggestions. The process to update the denylist is very manual and
developer heavy.

It might be difficult to update the denylist over the holiday period. In
that situation the decision might be taken to temporarily turn off
autocomplete entirely.

An environment variable is used to determine whether or not to render
search with autocomplete. The value of this variable will be applied
globally across all applications that use site search.

See https://github.com/alphagov/govuk-helm-charts/pull/2823