Fix test assertions that didn't actually assert
The `assert_no_selector` method is a method from the Capybara DSL that
only has an effect if you are navigating to a URL in the test and
populating a page attribute. In these tests there are is no page
attribute and thus no selectors to match against (just an empty string).

I verified these tests did nothing by checking them against
`assert_no_selector "*"` and they passed that.

As these are assertions that refute a state, these were likely not
noticed as these assertions would only fail if we visited a page that
had the selector.

I chose the approach of using `assert_select "selector", false` as this
seemed to be the most common refute selector in this repo with ~180 of
them compared to ~25 of `assert_select "selector", count: 0`. I used the
`count: 0` syntax when passing arguments other hash arguments as this
felt more idiomatic.

Annoyingly a new `assert_no_select` method will be available in the next
`rails-dom-testing` gem release (which provides `assert_select`) which would
be a better option than both of the above. [1]

[1]: https://github.com/rails/rails-dom-testing/blob/da0d96bc872e7e9b2656bc17ef3467bdf4bd9d48/CHANGELOG.md#next--unreleased